+++ src/main/java/kr/co/imguru/domain/pay/dto/PayCreateDto.java
@@ -1,6 +1,9 @@
 package kr.co.imguru.domain.pay.dto;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @Builder
@@ -8,20 +11,39 @@ import lombok.*;
 @NoArgsConstructor
 public class PayCreateDto {
 
-    private String resCode;
+//    private String hdiInfo;
 
-    private String trPriceEnc;
+//    private String apiVer;
 
-    private String emailEnc;
+//    private String processType;
 
-    private String cphoneNoEnc;
+//    private String dutyFreeYn;
+
+//    private String trTime;
+
+//    private String trDay;
+
+    private String mercntId;
+
+    private String ordNo;
+
+    private Long trPrice;
+
+    private String trPricePlain;
+
+    private String productNm;
+
+    private String mercntParam1;
+
+    private String mercntParam2;
+
+    private String viewType;
+
+    private String email;
+
+    private String cphoneNo;
 
     private String signature;
 
-    private String trDay;
-
-    private String trTime;
-
-    private String viewType;
 
 }
